requirements.txt
requests>=2.28.0
schedule>=1.1.0
PyYAML>=6.0
pandas>=1.5.0
